requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
cloudinary==1.41.0
crispy-bootstrap4==2024.10
dj-database-url==0.5.0
dj-rest-auth==7.0.0
dj3-cloudinary-storage==0.0.6
Django==4.2.17
django-allauth==65.3.1
django-cors-headers==4.6.0
django-crispy-forms==2.3
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.8.2
